+++ apps/web/src/lib/postbuild.ts
@@ -1,210 +1,16 @@
-import { existsSync, readFileSync, writeFileSync, readdirSync, statSync } from 'fs';
-import { join, resolve } from 'path';
-import { execSync } from 'child_process';
-import { Font, woff2 } from 'fonteditor-core';
-import { parse } from 'node-html-parser';
+import { existsSync } from 'fs';
+import { join } from 'path';
+import { restoreBlogRoutes } from './postbuild/restore-blog-routes';
+import { initFontSubsetting, injectCriticalFonts } from './postbuild/critical-fonts';
+import { injectRouteList } from './postbuild/preload-routes';
+import { findHtmlFiles } from './postbuild/utils';
 
-const WEB_DIR = process.cwd();
-const BUILD_DIR = join(WEB_DIR, 'build');
-const TEMP_DIR = join(WEB_DIR, '.temp');
-const FONTS_DIR = resolve(WEB_DIR, '../../packages/svelte/static/fonts');
-
-// ────────────────────────────────────────────
-// Blog route restoration
-// ────────────────────────────────────────────
-
-function restoreBlogRoutes() {
-  if (!existsSync(TEMP_DIR)) {
-    console.log('postbuild: No .temp directory found');
-    return;
-  }
-
-  const tempBlogDir = join(TEMP_DIR, 'src', 'routes', 'blog', '[title]');
-  const blogDir = join(WEB_DIR, 'src', 'routes', 'blog');
-
-  if (existsSync(tempBlogDir)) {
-    execSync(`mv "${tempBlogDir}" "${join(blogDir, '[title]')}"`, { stdio: 'inherit' });
-    console.log('postbuild: Restored blog routes from .temp directory');
-  }
-
-  execSync(`rm -rf "${TEMP_DIR}"`, { stdio: 'inherit' });
-  console.log('postbuild: Removed .temp directory');
-}
-
-// ────────────────────────────────────────────
-// Critical font injection
-// ────────────────────────────────────────────
-
-interface FontChars {
-  inter: Set<string>;
-  interItalic: Set<string>;
-  garamond: Set<string>;
-  garamondItalic: Set<string>;
-}
-
-/** Add visible characters (spaces + non-whitespace) from text into a set. */
-function addChars(set: Set<string>, text: string) {
-  for (const ch of text) {
-    if (ch === ' ' || ch.trim()) set.add(ch);
-  }
-}
-
-/** Query the DOM and collect characters grouped by font-family and style. */
-function collectChars(html: string): FontChars {
-  const chars: FontChars = {
-    inter: new Set(),
-    interItalic: new Set(),
-    garamond: new Set(),
-    garamondItalic: new Set()
-  };
-
-  const root = parse(html);
-  const body = root.querySelector('body');
-  if (!body) return chars;
-
-  // Garamond — all text inside .font-garamond elements
-  for (const el of body.querySelectorAll('.font-garamond')) {
-    addChars(chars.garamond, el.textContent);
-  }
-
-  // Garamond italic — italic text within garamond sections
-  for (const el of body.querySelectorAll('.font-garamond .italic')) {
-    addChars(chars.garamondItalic, el.textContent);
-  }
-
-  // Inter — everything except garamond, mono, script, style
-  const interClone = parse(body.outerHTML);
-  for (const el of interClone.querySelectorAll('.font-garamond, .font-mono, script, style')) {
-    el.remove();
-  }
-  addChars(chars.inter, interClone.textContent);
-
-  // Inter italic — italic text within inter sections
-  for (const el of body.querySelectorAll('.font-inter .italic')) {
-    addChars(chars.interItalic, el.textContent);
-  }
-
-  return chars;
-}
-
-function subsetToBase64(fontPath: string, chars: Set<string>): string | null {
-  if (chars.size === 0) return null;
-  const codepoints = [...chars].map((c) => c.codePointAt(0)!);
-  const fontBuffer = readFileSync(fontPath);
-  const font = Font.create(fontBuffer, { type: 'woff2', subset: codepoints });
-  const output = font.write({ type: 'woff2' }) as ArrayBuffer;
-  return Buffer.from(new Uint8Array(output)).toString('base64');
-}
-
-function unicodeRangeFromChars(chars: Set<string>): string {
-  const codepoints = [...chars].map((c) => c.codePointAt(0)!).sort((a, b) => a - b);
-
-  // Group consecutive codepoints into ranges: [97,98,99,120] → [[97,99],[120,120]]
-  const ranges: [number, number][] = [];
-  for (const cp of codepoints) {
-    const last = ranges.at(-1);
-    if (last && cp === last[1] + 1) {
-      last[1] = cp; // extend current range
-    } else {
-      ranges.push([cp, cp]); // start new range
-    }
-  }
-
-  return ranges
-    .map(([start, end]) =>
-      start === end
-        ? `U+${start.toString(16).toUpperCase()}`
-        : `U+${start.toString(16).toUpperCase()}-${end.toString(16).toUpperCase()}`
-    )
-    .join(',');
-}
-
-function injectCriticalFonts(htmlPath: string) {
-  const html = readFileSync(htmlPath, 'utf-8');
-  const chars = collectChars(html);
-
-  const fontFaces: string[] = [];
-  const fonts: { chars: Set<string>; file: string; family: string; style: string }[] = [
-    { chars: chars.inter, file: 'Inter.woff2', family: 'Inter', style: 'normal' },
-    { chars: chars.interItalic, file: 'Inter-Italic.woff2', family: 'Inter', style: 'italic' },
-    { chars: chars.garamond, file: 'EBGaramond.woff2', family: 'EB Garamond', style: 'normal' },
-    {
-      chars: chars.garamondItalic,
-      file: 'EBGaramond-Italic.woff2',
-      family: 'EB Garamond',
-      style: 'italic'
-    }
-  ];
-
-  for (const { chars: charSet, file, family, style } of fonts) {
-    if (charSet.size === 0) continue;
-    const b64 = subsetToBase64(join(FONTS_DIR, file), charSet);
-    if (!b64) continue;
-    const unicodeRange = unicodeRangeFromChars(charSet);
-    fontFaces.push(
-      `@font-face{font-family:'${family}';src:url(data:font/woff2;base64,${b64}) format('woff2');font-style:${style};font-display:block;unicode-range:${unicodeRange}}`
-    );
-  }
-
-  if (fontFaces.length === 0) return;
-
-  const styleTag = `<style data-critical-font>${fontFaces.join('')}</style>`;
-  const injected = html.replace('</head>', `${styleTag}</head>`);
-  writeFileSync(htmlPath, injected);
-
-  const totalKB = (Buffer.byteLength(styleTag) / 1024).toFixed(1);
-  const relativePath = htmlPath.replace(BUILD_DIR, '');
-  console.log(`criticalFonts: ${relativePath} — ${fontFaces.length} font(s), ${totalKB} KB inline`);
-}
-
-// ────────────────────────────────────────────
-// Route list injection (for client-side preloading)
-// ────────────────────────────────────────────
-
-function injectRouteList(htmlFiles: string[]) {
-  const allRoutes = htmlFiles.map((f) => {
-    const rel = f
-      .replace(BUILD_DIR, '')
-      .replace(/\/index\.html$/, '/')
-      .replace(/\.html$/, '');
-    return rel || '/';
-  });
-
-  const script = `<script>window.__ALL_ROUTES=${JSON.stringify(allRoutes)}</script>`;
-
-  for (const f of htmlFiles) {
-    const html = readFileSync(f, 'utf-8');
-    writeFileSync(f, html.replace('</head>', `${script}</head>`));
-  }
-
-  console.log(`postbuild: Injected ${allRoutes.length} routes for preloading`);
-}
-
-// ────────────────────────────────────────────
-// Helpers
-// ────────────────────────────────────────────
-
-function findHtmlFiles(dir: string): string[] {
-  const files: string[] = [];
-  for (const entry of readdirSync(dir)) {
-    const full = join(dir, entry);
-    if (statSync(full).isDirectory()) {
-      files.push(...findHtmlFiles(full));
-    } else if (entry.endsWith('.html')) {
-      files.push(full);
-    }
-  }
-  return files;
-}
-
-// ────────────────────────────────────────────
-// Main
-// ────────────────────────────────────────────
+const BUILD_DIR = join(process.cwd(), 'build');
 
 restoreBlogRoutes();
 
 if (existsSync(BUILD_DIR)) {
-  await woff2.init();
+  await initFontSubsetting();
   const htmlFiles = findHtmlFiles(BUILD_DIR);
 
   console.log(`postbuild: Processing ${htmlFiles.length} HTML files...`);
