+++ apps/web/src/lib/postbuild/utils.ts
@@ -1,0 +1,15 @@
+import { readdirSync, statSync } from 'fs';
+import { join } from 'path';
+
+export function findHtmlFiles(dir: string): string[] {
+  const files: string[] = [];
+  for (const entry of readdirSync(dir)) {
+    const full = join(dir, entry);
+    if (statSync(full).isDirectory()) {
+      files.push(...findHtmlFiles(full));
+    } else if (entry.endsWith('.html')) {
+      files.push(full);
+    }
+  }
+  return files;
+}
